Update to be compatible with node 16
This commit updates various dependencies to their latest versions and
updates a small amount of code to be compatible with the new versions.
Notably this updates node-sass and the collection of gulp-related
packages, both of which are prone to compatibility issues with
unsupported versions of nodejs.

This commit was not done artfully. I have very little experience with
the JS ecosystem, all I really did was follow error messages and stack
overflow answers until it ran successfully. A more experienced person
may want to review this changes with more scrutiny than I gave.